Split mixed API tests for crosscheck
Refs #212
