+++ safere/src/test/java/org/safere/InstTest.java
@@ -10,6 +10,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Test;
 
 /** Tests for {@link Inst}. */
+@DisabledForCrosscheck("implementation test uses package-private SafeRE internals")
 class InstTest {
 
   @Test
