+++ safere/src/test/java/org/safere/ParserTest.java
@@ -17,6 +17,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 /** Tests for {@link Parser}. */
+@DisabledForCrosscheck("implementation test uses package-private SafeRE internals")
 class ParserTest {
 
   /** Perl-compatible flags, used as the default for most tests. */
