+++ safere/src/test/java/org/safere/ProgTest.java
@@ -12,6 +12,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Test;
 
 /** Tests for {@link Prog}. */
+@DisabledForCrosscheck("implementation test uses package-private SafeRE internals")
 class ProgTest {
 
   @Test
